09.19

+++ Assets/02.Scripts/All Inventory/Item Inventory/ItemData/PortionItemData.cs
@@ -13,12 +13,12 @@ public class PortionItemData : CountableItemData
 {
     ///</summary> 효과랑(회복량 등) </summary>
     [Newtonsoft.Json.JsonProperty]
-    private float _value;
+    private int _value;
 
     [Newtonsoft.Json.JsonProperty]
     private IPortionState _portionState;
 
-    public float GetValue() { return _value; }
+    public int GetValue() { return _value; }
     public IPortionState GetPortionState() { return _portionState; }
 
     public override Item CreateItem()
